services/slav/app/requirements.txt
grpcio==1.33.1
grpcio-reflection==1.33.1
psycopg2==2.8.5
schedule==0.6.0
SQLAlchemy==1.3.18
